requirements.txt
av
decord
einops
fathomnet
ftfy
huggingface-hub
hydra-core
iopath
matplotlib
numba
numpy==1.26.4
omegaconf
openai
opencv-python-headless
pandas
Pillow
psutil
pycocotools
pytest
PyYAML
regex
requests
scikit-learn
scipy
setuptools
scikit-image
submitit
tensordict
tidecv
timm
torch
torchcodec
torchmetrics
torchvision
tqdm
typing-extensions
ultralytics
yt-dlp